:recycle: refactor: 홈페이지 컴포넌트 구조 개선
